health/views.py
from rest_framework.views import APIView
from rest_framework.response import Response

# This class likely represents a view in a Python web application that performs health checks.
class HealthCheck(APIView):
    def get(self, requet):
        # The line `return Response({"status": "ok"}, status=200)` is returning a response to the
        # client with a JSON object `{"status": "ok"}` and an HTTP status code of 200 (OK). This is a
        # common pattern in web development where an API endpoint responds with a JSON payload to
        # indicate the health status of the application.
        return Response({"status": "ok"}, status=200)