pyproject.toml
[tool.poetry]
name = "hypertube"
version = "0.1.0"
description = ""
authors = ["kmlFaouzi <faouzikamal27@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
dj-database-url = "^2.2.0"
python-decouple = "^3.8"
django-environ = "^0.11.2"
django = "^5.1.2"
djangorestframework = "^3.15.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
pre-commit = "^3.8.0"
importlib-metadata = "4.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
